bpf: Bypass host fw for kube-proxy SNATed connections
[ upstream commit dc6a0655172d3750f581db52295c4588d1182e18 ]

If kube-proxy is in use (no BPF-based masquerading), packets from pods may
be SNATed. The response packet will therefore have a host IP as the
destination IP. For example, if a pod connects to 1.1.1.1:

  # Pod sends a SYN.
  from-lxc: 10.8.0.1 -> 1.1.1.1
  # Src IP gets masqueraded to 192.168.33.11. We skip the host firewall
  # because the packet mark tells us this is not coming from hostns.
  to-netdev: 192.168.33.11 -> 1.1.1.1
  ...
  # SYN+ACK is sent to 192.168.33.11. This IP belongs to the host
  # endpoints, so we incorrectly enforce host policy on this packet.
  from-netdev: 1.1.1.1 -> 192.168.33.11

To avoid enforcing host policies for response packets to pods, we need to
create a CT entry for the forward, SNATed packet from the pod. Response
packets will thus match this CT entry and bypass host policies.

I tested this change by deploying the connectivity checks in the dev. VM
with kube-proxy and the lock-down-dev-vms.yaml host policy. A subsequent
commit also updates the CIDR+toPorts host policy test to run it when
kube-proxy is used. Finally, the pull request has label ci/host-firewall
to run all e2e CI tests on Jenkins with the host firewall enabled by
default.

Fixes: 88bf291 ("bpf: Enforce host policies for IPv4")
Signed-off-by: Paul Chaignon <paul@cilium.io>
Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>